+++ scripts/startUserMigration.js
@@ -1,13 +1,14 @@
 const { program } = require('commander');
 const appPromise = require('../src/app');
-const { error } = require('../src/logger');
+const { alert, error } = require('../src/logger');
 const { schoolModel } = require('../src/services/school/model');
+const systemModel = require('../src/services/system/model');
 
 program
-	.option('-s, --systemId <value>', 'id of the system, which will be added to the schools (required argument)')
+	.requiredOption('--systemId <value>', 'ID of the system, which will be added to the schools')
 	.option(
-		'-S, --schoolIds <value...>',
-		'ids of the schools that should be migrated. If not given all schools with a official school number will be migrated.',
+		'--officialSchoolNumbers <value...>',
+		'Official school numbers of the schools that should be migrated (space seperated values). If not given, all schools having an official school number will be migrated.'
 	);
 
 program.parse();
@@ -15,17 +16,31 @@ program.parse();
 const options = program.opts();
 
 const startUserMigration = async () => {
+	alert('user migration script started');
 	const { systemId } = options;
-	const schoolsToMigrate = options.schoolIds;
+	const system = await systemModel.findOne({ _id: systemId, 'ldapConfig.rootPath': { $exists: true } });
+	if (!system) {
+		throw new Error(`System with the id ${systemId} cannot be found. Provide a valid system id instead.`);
+	}
+	const schoolsToMigrate = options.officialSchoolNumbers;
 	let schools;
 	if (!schoolsToMigrate) {
 		schools = schoolModel.find({ officialSchoolNumber: { $exists: true } });
+		const numberOfFoundSchools = await schools.count();
+		alert(`${numberOfFoundSchools} schools found that will be migrated`);
 	} else {
-		schools = schoolModel.find({ _id: { $in: schoolsToMigrate } });
+		schools = schoolModel.find({ officialSchoolNumber: { $in: schoolsToMigrate } });
+		const numberOfFoundSchools = await schools.count();
+		if (numberOfFoundSchools !== schoolsToMigrate.length) {
+			throw new Error(
+				`From the given list of schools to migrate, not all schools could be found in the database. Check the provided schools and execute the script again afterwards.`
+			);
+		}
+		alert(`${numberOfFoundSchools} schools found that will be migrated`);
 	}
 	for await (const school of schools) {
 		try {
-			school.ldapSchoolIdentifier = school.officialSchoolNumber;
+			school.ldapSchoolIdentifier = `ou=${school.officialSchoolNumber},${system.ldapConfig.rootPath}`;
 			if (!school.systems.includes(systemId)) school.systems.push(systemId);
 			school.inUserMigration = true;
 			school.inMaintenanceSince = new Date();
@@ -35,6 +50,7 @@ const startUserMigration = async () => {
 			error(`school ${school._id} cannot be set to user migration mode`, err);
 		}
 	}
+	alert('user migration script finished');
 	return process.exit(0);
 };