+++ CHANGELOG.md
@@ -11,7 +11,7 @@ Allowed Types of change: `Added`, `Changed`, `Deprecated`, `Removed`, `Fixed`, `
 
 ### Added
 
-- BC-837 - BC-932 - add script to prepare ldap migration for schools
+- BC-837 - BC-932 - add script to start ldap migration for schools
 
 ### Changed